Move test classes to the packages of their domain class - (Task #122)
