+++ de.dlr.sc.virsat.requirements.tracing.test/src/de/dlr/sc/virsat/requirements/tracing/util/TraceHelperTestExtension.java
@@ -10,9 +10,7 @@
 /**
  * 
  */
-package de.dlr.sc.virsat.requirements.tracing.util.test;
-
-import de.dlr.sc.virsat.requirements.tracing.util.TraceHelper;
+package de.dlr.sc.virsat.requirements.tracing.util;
 
 /**
  * @author fran_tb
