+++ de.dlr.sc.virsat.requirements.tracing.validation.test/src/de/dlr/sc/virsat/requirements/tracing/validation/test/AllTests.java
@@ -13,6 +13,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
+import de.dlr.sc.virsat.requirements.tracing.validation.engines.InspectionWarningEngineTest;
 import junit.framework.JUnit4TestAdapter;
 
 /**
